+++ tests/elisions.rs
@@ -3,7 +3,7 @@
 #![warn(rust_2018_idioms, single_use_lifetimes)]
 #![allow(clippy::needless_pass_by_value)]
 #![allow(clippy::needless_lifetimes)] // broken
-#![feature(generators)]
+#![feature(coroutines)]
 
 use futures_async_stream::stream;