+++ tests/no-std/lib.rs
@@ -2,7 +2,7 @@
 
 #![no_std]
 #![warn(rust_2018_idioms, single_use_lifetimes)]
-#![feature(generators)]
+#![feature(coroutines)]
 
 use futures_async_stream::{stream, try_stream};