+++ tests/overwriting_core_crate.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 #![warn(rust_2018_idioms, single_use_lifetimes)]
-#![feature(generators)]
+#![feature(coroutines)]
 
 // See https://github.com/rust-lang/pin-utils/pull/26#discussion_r344491597
 //