+++ tests/stream.rs
@@ -3,7 +3,7 @@
 #![warn(rust_2018_idioms, single_use_lifetimes)]
 #![allow(clippy::unused_async)]
 #![allow(clippy::needless_lifetimes, clippy::semicolon_if_nothing_returned)] // broken
-#![feature(generators, proc_macro_hygiene, stmt_expr_attributes, gen_future)]
+#![feature(coroutines, proc_macro_hygiene, stmt_expr_attributes, gen_future)]
 
 use std::{pin::Pin, rc::Rc, sync::Arc};